requirements.txt
streamlit
python-dotenv
Pillow
pdf2image
google-generativeai
python-docx
scikit-learn
setup.py
setup.cfg
Pipfile 